refactor: extract mutation hook
