--- src/data-workspace/use-mutation-fn.js
@@ -1,7 +1,0 @@
-import { useDataEngine } from '@dhis2/app-runtime'
-
-export const useMutationFn = (mutation) => {
-    const engine = useDataEngine()
-
-    return (variables) => engine.mutate(mutation, { variables })
-}
